Handle Norman types via Steve
